requirements.txt
# Core
pydantic>=2.0
pyyaml>=6.0

# PDF processing
pymupdf>=1.24.0
Pillow>=10.0
pdf2image>=1.17.0  # fallback if pymupdf unavailable

# Embeddings
colpali-engine>=0.4.0
torch>=2.0
fast-plaid  # optional: accelerates ColQwen2 search

# Vector store
qdrant-client>=1.12.0

# LLM
anthropic>=0.42.0

# API server
fastapi>=0.115.0
uvicorn>=0.30.0

# Logging
structlog>=24.0

# Testing
pytest>=8.0
pytest-asyncio>=0.24.0
